Tighten path normalization and Iceberg integer widths
ManifestListEntry.sequenceNumber and min_sequence_number were typed as
Int while the Iceberg spec defines them as int64; tables with > 2³¹
sequence numbers would silently overflow during Avro decode. The path
normalizer dropped the authority component of file://host/share/path
URIs, mapping them to the wrong local path. parsePosition truncated
Long → Int instead of rejecting out-of-range values.

- IcebergSchema: sequenceNumber / minSequenceNumber → Long?.
- IcebergGraphBuilder.unifiedManifestComparator: Int.MAX_VALUE → Long.MAX_VALUE.
- IcebergPaths.normalizeFilePath: reconstruct file://host/path as //host/path
  (UNC-style); empty / localhost authority unchanged.
- GraphLayoutService.parsePosition: reject values outside Int range.
- Avro test fixtures updated to write Long sequence numbers.

+++ src/main/kotlin/model/IcebergSchema.kt
@@ -113,8 +113,8 @@ data class ManifestListEntry(
     @SerialName("manifest_length") val manifestLength: Long? = null,
     @SerialName("partition_spec_id") val partitionSpecId: Int? = null,
     @SerialName("content") val content: Int? = null, // 0=Data, 1=Deletes
-    @SerialName("sequence_number") val sequenceNumber: Int? = null,
-    @SerialName("min_sequence_number") val minSequenceNumber: Int? = null,
+    @SerialName("sequence_number") val sequenceNumber: Long? = null,
+    @SerialName("min_sequence_number") val minSequenceNumber: Long? = null,
     @SerialName("added_snapshot_id") val addedSnapshotId: Long? = null,
     @SerialName("added_files_count") val addedFilesCount: Int? = null,
     @SerialName("existing_files_count") val existingFilesCount: Int? = null,
